requirements.txt
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
deep-sort-realtime==1.3.2
deepsort==0.0.5
dnspython==2.5.0
easyocr==1.7.1
filelock==3.13.1
filterpy==1.4.5
fonttools==4.47.0
fsspec==2023.12.2
h11==0.14.0
idna==3.6
imageio==2.33.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.25.2
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
packaging==23.2
panda==0.3.1
pandas==2.0.2
pillow==10.2.0
psutil==5.9.7
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pymongo==4.6.1
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.13.1
shapely==2.0.2
simple-websocket==1.0.0
six==1.16.0
supervision==0.17.1
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.0.236
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
